Moving card-validator into the src/payment/checkout directory
It’s only used there, so it makes sense for it to live with the code.
