+++ .eslintignore
@@ -2,4 +2,4 @@ coverage/*
 dist/
 node_modules/
 __mocks__/
-src/common/card-validator/
+src/payment/checkout/card-validator/
