+++ src/payment/checkout/card-validator/src/credit-card-type/LICENSE
